feat: Add models and seeders for Trip, Driver, Rating, Vehicle, and Passenger
This commit adds the necessary models and seeders for the Trip, Driver, Rating, Vehicle, and Passenger entities in the application. These models will be used to interact with the corresponding database tables and the seeders will populate the tables with initial data.

The models are created under the `App\Models` namespace and extend the `Illuminate\Database\Eloquent\Model` class. They also use the `HasFactory` trait for easy factory and seeding setup.

The seeders are created under the `Database\Seeders` namespace and extend the `Illuminate\Database\Seeder` class. They include a `run` method that will be used to populate the corresponding tables with data.

This commit sets up the foundation for working with the Trip, Driver, Rating, Vehicle, and Passenger entities in the application.