feat: Add default attribute to Vehicle and is_available attribute to Driver
This commit adds the `default` attribute to the `Vehicle` model and the `is_available` attribute to the `Driver` model. The `default` attribute in the `Vehicle` model is used to indicate if a vehicle is the default vehicle for a driver. The `is_available` attribute in the `Driver` model is used to indicate if a driver is currently available for assignments. These attributes provide additional functionality and flexibility to the application.